refactor: rename recorder persistence module
